backend/requirements.txt
fastapi
uvicorn
langgraph
langchain
langchain-openai
openai
pydantic
python-multipart
requests
python-dotenv
aiofiles
aiohttp
psutil
langfuse
